refactor: move the definitions of CDB to the proper module
